move camera feed specific hooks to CameraFeed dir
